[#54] Adjust main.load_conf() to set a "log_dir" attribute instead of a "log_file" attribute
This lets us decide what to name the log file later in main._setup_logging(). This will
be helpful when we want to change the name of the log for each hlink run.

This also slightly changes the current logging setup. It creates a logs/ subdirectory
and outputs logs to an hlink.log file in that subdirectory.