[#53, #38] Move lots of package metadata into pyproject.toml
This moves lots of metadata from setup.cfg and setup.py into pyproject.toml.
By using the setuptools-scm plugin, we can get rid of lots of the gnarly
package data logic in setup.py, which is great. This does make use of the
beta tool.setuptools table in pyproject.toml. So we should keep an eye on that.

setup.py is still around for backwards compatibility, but we've done away
with setup.cfg completely.