Drop the comment column in the desc command
This column is automatically returned when running 'DESCRIBE table', but we don't
really use it in hlink at the moment. So it's all nulls. This is a bug; we should
just not show it to the user. The desc command gives you the names of the columns
and their data types in Spark.