[#80] Deprecate the hlink.linking.transformers.interaction_transformer module
This module contains a single class named InteractionTransformer which was
originally included as a backport of the pyspark.ml.feature.Interaction class to
pyspark 2. Now that we're on pyspark 3, we can just use the Interaction class
instead, and users can as well. So we can deprecate this module to prevent drift
between our backport and the pyspark implementation.

Removing this class is a breaking change, so we will need to wait until hlink 4
to do that.