Add the necessary columns to exploded_df_a and exploded_df_b for multi_jaro_winkler_search
This happens automatically for comparison_features that have the column_name or column_names
attributes. But multi_jaro_winkler_search uses templating, so it doesn't have these
attributes. So we need a special case to make sure that the columns needed for the comparison
feature are present in the exploded_df_a and exploded_df_b when we try to run the SQL query.