[#105] Don't pull training data into Pandas
This could potentially be a pretty big dataset, so let's avoid pulling it into
memory with Pandas. Instead, we can pull a single row into memory with .head()
and analyze that to get the number of categories.

I've also added strict=True to the zip() to check that the number of feature names
that we generate is exactly the same as the number of coefficients found.