[#110] Add a failing test for the interacted categorical feature coefficients bug
This bug is triggered when a categorical feature is interacted with another feature.
The resulting feature is also categorical, but its name may not end with _onehotencoded
because the user can specify its name in the config file. So we can't rely on
column names ending with _onehotencoded to specify which features are categorical.