[#110] Fix a bug where interacted categorical features caused Training step 3 to crash
These features are categorical but may not end with the suffix "_onehotencoded". We now
check the data type of the columns in training_features_prepped and take any column with
type "vector" to be categorical and explode it. This seems pretty robust.

I also improved the test for this bug a bit.