Upgrade a bunch of dependencies
Mostly we've just removed the upper bounds on a lot of packages we depend upon.
These seem unlikely to break things, and we have a bunch of tests to detect when
they might. pyspark is so central to what we do that I've left it ~-pinned for
now. I've also done the same upgrades for some development dependencies. I'll
need to rerun black for the new version.