[#142] Rework _explode() to support multiple exploded columns
Instead of trying to select() out all of the columns we need in each iteration,
we can use withColumn() to add the exploded columns one by one. There is a
weird bit of logic where we need to do an extra select only if there are
exploded columns. I added a comment about that and will do a bit more looking
into it.