[#162] Add a rough draft of a RenameVectorAttributes transformer
Usually we don't care about the names of the vector attributes. But LightGBM
uses them as feature names and disallows some characters in the names.
Unfortunately, one of these characters is :, and Spark's Interaction names the
output of an interaction between A and B "A:B". I looked through the Spark code
and didn't see any way to configure the names of these output features. So I
think the easiest way forward here is to make a transformer that renames the
attributes of a vector by removing some characters and replacing them with
another.