[#199] Compute threshold ratio even if the second best probability is below alpha threshold
If I just compute the threshold ratio for every pair with a second best
probability, one of the tests fails. So I've set it up to only compute the
threshold ratio if the best probability is at least the alpha threshold. I'll
look into this, since I think it would be simpler to always compute the
threshold ratio when there's a second best probability.