[#205] Refactor the big if statemet to a dict
This is a bit cleaner and more compact. It should make it easier to add new
built in column mapping transforms, since you won'thave to mess around with all
of the if branches. But the big improvement with this is that it makes it clear
that we could let users define their own custom column mapping transforms.
Adding a custom column mapping transform is conceptually just adding a new
key-value entry to the map with the transform type name and the function which
computes the transform expression. In practice we may do something a little bit
different, since we have some builtin transform types which should always be
available, etc.