Upgrade to Spark 4 and Scala 2.13 (still a work in progress)
Spark 4 requires Scala 2.13, so we have to upgrade. Spark 4 sets the
spark.sql.ansi.enabled config option to true by default, which changes
the behavior of many casts and math operations. I've reset it to false
for now to retain the previous behavior.

LightGBM does not yet support Spark 4 from what I can tell. So it is
broken.